* Fixed: properly compare User/Zone IDs in User/Zone Rule casting them to strings

* Expire Price List cache key when adding/removing products

* Include Price List cache key in `spree_storefront_base_cache_key`

+++ core/app/models/spree/price_list.rb
@@ -155,6 +155,7 @@ module Spree
       # Use upsert_all with on_duplicate: :skip to handle race conditions
       Spree::Price.upsert_all(prices_to_insert, on_duplicate: :skip)
       touch_variants(variant_ids)
+      touch
     end
 
     # Removes products from the price list
@@ -165,10 +166,13 @@ module Spree
       return if product_ids.blank?
 
       variant_ids = Spree::Variant.where(product_id: product_ids).distinct.pluck(:id)
+      return if variant_ids.empty?
+
       # Use delete_all for hard delete - this bypasses acts_as_paranoid
       # which is required for the unique index to work when re-adding products
       prices.where(variant_id: variant_ids).delete_all
       touch_variants(variant_ids)
+      touch
     end
 
     # Bulk update prices using upsert_all for performance
